Listagem de cidades por estado e criaçao dos objetos a partir dos campos dos inputs

+++ backend/FormsDev/db/migrate/20210519180311_create_enderecos.rb
@@ -3,7 +3,7 @@ class CreateEnderecos < ActiveRecord::Migration[6.1]
     create_table :enderecos do |t|
       t.string :logradouro
       t.string :endereco
-      t.numeric :numero
+      t.integer :numero
       t.string :cep
 
       t.timestamps
